Split chainstate/storage/internal module
